Add support for nested userdata in generic authenticator.
Services like NextCloud return a nested data-structure at the userdata
endpoint. A single key to get the username is not sufficient here.
This adds a new envvar "OAUTH2_USERNAME_KEY_NESTED" to tell that the
username-key is a dot-separated sequence of keys to access nested
userdata responses. For Nextcloud this would be "ocs.data.id".

Unittests are implemented to test also with nested data.