PyJWT 2.0 compliant
PyJWT deprecated the argument `verify` and made it part of options as `verify_signature`. This PR solves that bug, basically the same as this one: https://github.com/Clinical-Genomics/cgbeacon2/issues/116

you will get:
jwt.exceptions.DecodeError: It is required that you pass in a value for the "algorithms" argument when calling decode().