Add CanvasOAuthenticator
Canvas is a very popular educational tool used to manage
classes in universities across the world. It has a very
rich API (https://canvas.instructure.com/doc/api/), and
deep integration with it makes a lot of things possible.

This Authenticator builds on top of GenericOAuthenticator,
but adds a few new features:

- Strip domain names specific user emails to get userids. Works
  when all your Canvas users have an email for your educational
  institution.
- Fetch list of courses the user is enrolled in, and what kind
  (student, TA, instructor, etc). This allows for finer
  grained access control, and modifying the user environment
  based on the courses they are enrolled in.
- Ability to restrict user logins based on what courses they
  are enrolled in.

I've added some unit tests, and things seem to be working out
ok.