Get rid of the okpy authenticator
There is exactly one installation of okpy in the world - at
https://okpy.org/. It's primarily used only by UC Berkeley,
and this authenticator was used for a short period of time at
one of the hubs run by UC Berkeley. However, it has *not* been
used there for many years, and there are no ongoing plans for it
to be used ever again. I can speak reasonably authoritatively
on this, given I helped run those hubs from inception until
very recently.

Given there are not really any other okpy installations in the
world, I think it's actually safe to remove this one.