Move username_claim being callable to oauth2 autheneticator
While trying to use Auth0 for authentication in one of our
hubs, we discovered that the most useful username_claim (`sub`)
produces usernames that look like
`oauth2|cilogon|http://cilogon.org/servera/users/43431`
(when using auth0 with CILogon). The last part of `sub` is
generally whatever is passed on to auth0, so it's going to
be different for different users.

I had thought `username_claim` was a callable, but turns out
that's only true for GenericOAuthenticator. I think it's
pretty useful for every authenticator, so I've just moved
that functionality out to the base class instead. I also
added a test to verify it works. The test is in GenericOAuthenticator
because it was the easiest place to put it, but it works
across authenticators. This also means it is fully backwards
compatible.