Allow gating access based on scopes granted
We can already control what scopes we ask for by setting
'scope'. However, OAuth2
[states](https://datatracker.ietf.org/doc/html/rfc6749#section-3.3)
that not all the scopes requested may be granted, for whatever reason.
This could be because the user choose not to give us the grant,
or because the user themselves doesn't have access to grant us this
scope (This is how Auth0 uses it for example -
https://auth0.com/docs/get-started/authentication-and-authorization-flow/authorization-code-flow/call-your-api-using-the-authorization-code-flow).

This PR adds a `required_scopes` property, listing all the scopes
that *must* be granted for the user to be allowed access. By default
it's empty, so no behavior is changed.