Suggest possible config fix when correct scopes aren't requested
If the hub config doesn't contain appropriate scopes to request
from the provider, the token we retrieve may not have the key we
are using as `username_claim`. This is the *most common* reason that
`username_claim` is not present in the token. Add a simple debugging
hint here to make lives of hub admins easier.