Test id token
Note, the current mock setup delivers different handlers (for ingesting
sample users) depending on whether id tokens are intended to be supplied.
The fixtures (for preparing a mock client and preconfigured authenticator)
thus need specialisation depending on whether the id token is being used.
However, the current mock also requires that tests individually manage the
packaging of the id token when loading into the mock client's handler;
the handler syntax is different between the two cases. This means that to
test that JWT id tokens and JSON userinfo endpoints provide equivalent
behaviour requires changes in both the test (in the preamble) and in
fixtures (compared to unmodified tests). This has led to some code
duplication, which would be a target for future refactoring.